+++ test/init.test.ts
@@ -14,8 +14,6 @@ it("supports --help", async ({ expect, prismic }) => {
 	const { stdout, stderr, exitCode } = await prismic("init", ["--help"]);
 	expect(exitCode, stderr).toBe(0);
 	expect(stdout).toContain("prismic init [options]");
-	expect(stdout).toContain("SUBCOMMANDS");
-	expect(stdout).toContain("instant");
 });
 
 it("fails if prismic.config.json already exists", async ({ expect, prismic }) => {
